+++ apps/clients/views.py
@@ -1,7 +1,7 @@
 from django.core.paginator import Paginator
 from django.shortcuts import get_object_or_404, redirect, render
 
-from .forms.form import ClientForm
+from .forms.clients_form import ClientForm
 from .models import Client
