+++ apps/goods_receipts/templates/pages/GRindex.html
@@ -1,7 +1,7 @@
 {% extends "layouts/base.html" %}{% block content %}{% load static %}
 
 <div class="flex justify-end mt-20">
-  <a href="{% url 'goods_receipts:GRnew' %}" class="btn bg-orange-600 text-white"
+  <a href="{% url 'goods_receipts:new' %}" class="btn bg-orange-600 text-white"
     >Add</a
   >
 </div>
